requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-MySQLdb==2.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
fonttools==4.58.5
fqdn==1.5.1
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
murmurhash==1.0.13
mysql-connector-python==9.3.0
mysqlclient==2.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==25.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.7.2
PyMySQL==1.1.1
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
qiskit==1.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
rustworkx==0.16.0
safetensors==0.5.3
scikit-learn==1.4.2
scipy==1.15.3
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
stevedore==5.4.1
symengine==0.14.1
sympy==1.12
terminado==0.18.1
textblob==0.19.0
thinc==8.2.5
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
typer==0.9.4
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.17.2
WTForms==3.1.2
Flask==2.3.2